Reimplement some of the UDP changes we have. Change how we send objects, this now puts them back in the queue, as we can't be sure that we should be sending them at the moment the thread sends them. Now that tokens are flowing through the parent buckets from the children, this should not be such a problem.

+++ OpenSim/Framework/Physics/ISceneViewer.cs
@@ -60,7 +60,7 @@ namespace OpenSim.Framework
         /// It loops through the available updates and sends them out (no waiting)
         /// </summary>
         /// <param name="numUpdates">The number of updates to send</param>
-        void SendPrimUpdates(int numUpdates);
+        void SendPrimUpdates(Int64 numUpdates);
 
         /// <summary>
         /// The client has left this region and went into a child region, clean up anything required
